- Se agregaron dos nuevos metodos a la clase LibrosBLL (SumarSalidaLibros y RestarSalidaLibros)
- Se hicieron algunos cambios en las siguientes clases; (DevolucionesBLL, PrestamosBLL, UsuariosBLL, SalidasLibrosBLL y todas las clases de la carpeta Entidades)

- Ahora cuando entra o sale algun libro, se actualiza la existencia del libro

- Se hicieron cambios en los registros;(Devoluciones, EntradasLibros, SalidasLibros y Prestamos)

+++ Entidades/DevolucionesDetalle.cs
@@ -15,7 +15,7 @@ namespace ProyectoFinal_PrestamosLibros.Entidades
         public int DevolucionId { get; set; }
         public int LibroId { get; set; }
         public double LibrosDevueltos { get; set; }
-        public double Dias { get; set; }
+        public int Dias { get; set; }
 
         //———————————————————————————[ ForeingKey ]———————————————————————————
         [ForeignKey("LibroId")]
